+++ components/predict-form/question-types/MultiChoiceQuestion.tsx
@@ -2,10 +2,10 @@ import { PlusIcon } from "@heroicons/react/20/solid"
 import { QuestionType } from "@prisma/client"
 import { useCallback, useState } from "react"
 import { useFormContext } from "react-hook-form"
-import { QuestionOption } from "../../questions/QuestionOption"
 import { FormCheckbox } from "../../ui/FormCheckbox"
 import { EmbeddedOptions } from "../EmbeddedOptions"
 import { PredictButton } from "../PredictButton"
+import { QuestionOption } from "../QuestionOption"
 import { ResolveBy } from "../ResolveBy"
 import { QuestionTypeProps } from "./question-types"
